+++ projects/ui-framework/src/lib/form-elements/radio-button/radio-button.enum.ts
@@ -1,0 +1,4 @@
+export enum RadioDirection {
+  row = 'row',
+  column = 'column'
+}
