+++ projects/ui-framework/src/lib/form-elements/radio-button/radio-button.interface.ts
@@ -1,0 +1,4 @@
+export interface RadioConfig {
+  id: number;
+  label: string;
+}
